fix: resolve 6 confirmed bugs from agent review
- Bug 1: insert_before/insert_after now auto-append/prepend newline
  separator using document's native line ending (LF or CRLF),
  preventing block content from being joined to adjacent lines
- Bug 3: search preview is now match-centred (±80 chars) so
  match_ranges offsets always fall within the preview string
- Bug 5: MCP ZodError now reports only the first failing field path
  and message instead of dumping the entire schema
- Bug 6: OffsetMapper switched from Int32Array to Float64Array to
  avoid silent integer overflow on files > 2 GB
- Bug 7: anchor token TTL raised to 1 hour (was 10 min) and is now
  configurable via MDSE_TOKEN_TTL_MS environment variable
- Bug 8: formatError unknown-code branch includes the code name in
  the recommended_action string for easier debugging

Bugs 2 and 4 (delete list marker / ordered-list renumbering) are
intentional protocol-layer design decisions: byte-range operations
do not perform semantic Markdown restructuring. This is documented
in AGENTS.md safety invariant #11.

+++ packages/core/src/snapshot/snapshot.ts
@@ -136,15 +136,17 @@ export function createSnapshot(targetPath: string, allowedRoots: string[]): Docu
  * Helper class to map between JS character offsets and UTF-8 byte offsets.
  */
 export class OffsetMapper {
-  private charToByte: Int32Array;
-  private byteToChar: Int32Array;
+  private charToByte: Float64Array;
+  private byteToChar: Float64Array;
 
   constructor(bytes: Uint8Array, str: string) {
     // Account for potential BOM offset in raw bytes relative to decoded string
     const bomOffset = bytes.length >= 3 && bytes[0] === 0xef && bytes[1] === 0xbb && bytes[2] === 0xbf ? 3 : 0;
     
-    this.charToByte = new Int32Array(str.length + 1);
-    this.byteToChar = new Int32Array(bytes.length + 1 - bomOffset);
+    // Float64Array is used instead of Int32Array to avoid silent integer overflow
+    // on very large files (Int32Array max ~2.1 GB). Float64 safely handles up to 2^53 bytes.
+    this.charToByte = new Float64Array(str.length + 1);
+    this.byteToChar = new Float64Array(bytes.length + 1 - bomOffset);
 
     let charIdx = 0;
     let byteIdx = 0;
@@ -183,7 +185,7 @@ export class OffsetMapper {
         this.charToByte[i] += bomOffset;
       }
       // Re-initialize byteToChar to include BOM
-      const newByteToChar = new Int32Array(bytes.length + 1);
+      const newByteToChar = new Float64Array(bytes.length + 1);
       for (let b = 0; b < bomOffset; b++) {
         newByteToChar[b] = 0;
       }
